+++ .gitignore
@@ -15,6 +15,7 @@ build/
 out/
 !**/src/main/**/out/
 !**/src/test/**/out/
+.idea/
 
 ### Eclipse ###
 .apt_generated
@@ -40,3 +41,4 @@ bin/
 
 ### Mac OS ###
 .DS_Store
+/.idea/
