+++ src/main/java/ru/otus/game/interfaces/IRotable.java
@@ -3,6 +3,5 @@ package ru.otus.game.interfaces;
 public interface IRotable {
     double getAngular();
     void setAngular(double newValue);
-
     double getAngularVelocity();
 }
